+++ ComplimentApp/app/src/main/java/com/matikano/complimentapp/presentation/compliment/ComplimentEvent.kt
@@ -1,0 +1,6 @@
+package com.matikano.complimentapp.presentation.compliment
+
+sealed class ComplimentEvent {
+    data class OnLoadCompliment(val content: String?): ComplimentEvent()
+    object OnRefresh: ComplimentEvent()
+}