+++ ComplimentApp/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.1.1] - 2022-10-17
+
+### Added
+* ComplimentEvent sealed class for handling different types of events in ComplimentScreen
+
+### Changed
+* Bug fixed double change of Compliment background while executing onRefresh function in SwipeRefresh composable
+* Code cleanup
+
+
 ## [0.1.0] - 2022-10-17
 
 ### Added